+++ config/kit.php
@@ -15,7 +15,7 @@ return [
     | contra a árvore de trabalho, que é mais ruidosa.
     */
 
-    'version' => '0.6.1',
+    'version' => '0.7.0',
 
     /*
     |--------------------------------------------------------------------------